requirements.txt
fastapi
uvicorn
streamlit
groq
langgraph
langchain
langchain-groq
chromadb
python-dotenv
pydantic
requests
feedparser